addons/ev_portal/rootfs/app/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
paho-mqtt==2.1.0
httpx==0.28.1
python-multipart==0.0.20
cryptography>=42.0.0
